requirements.txt
python-jobspy
fastmcp
mcp
fastapi
uvicorn
